Update jackson alignment rule for 2.20
There was a change in 2.20.x for the `jackson-annotations`
library. It no longer includes the patch version ([#294]).
Add it to the exclude list to avoid trying to align with
the patch versions of other jackson projects. If the other
projects are aligned, `jackson-annotations` should be
handled correctly with the normal dependency resolution.

[#294]: https://github.com/FasterXML/jackson-annotations/issues/294